* commit '03836b040d3bf53a93b8577060a4eb0c11639a5e':
  Switch the dimensions array to use uint32_t instead of size_t.

+++ cpp/rsDispatch.h
@@ -74,7 +74,7 @@ typedef void (*ScriptSetVarFFnPtr) (RsContext, RsScript, uint32_t, float);
 typedef void (*ScriptSetVarDFnPtr) (RsContext, RsScript, uint32_t, double);
 typedef void (*ScriptSetVarVFnPtr) (RsContext, RsScript, uint32_t, const void*, size_t);
 typedef void (*ScriptGetVarVFnPtr) (RsContext, RsScript, uint32_t, void*, size_t);
-typedef void (*ScriptSetVarVEFnPtr) (RsContext, RsScript, uint32_t, const void*, size_t, RsElement, const size_t*, size_t);
+typedef void (*ScriptSetVarVEFnPtr) (RsContext, RsScript, uint32_t, const void*, size_t, RsElement, const uint32_t*, size_t);
 typedef RsScript (*ScriptCCreateFnPtr) (RsContext, const char*, size_t, const char*, size_t, const char*, size_t);
 typedef RsScript (*ScriptIntrinsicCreateFnPtr) (RsContext, uint32_t id, RsElement);
 typedef RsScriptKernelID (*ScriptKernelIDCreateFnPtr) (RsContext, RsScript, int, int);
